Introduce 'fastClientDisconnectDetect' option
If the client publishes faster than the keepalive interval, PINGREQ messages keep getting rescheduled and will not be sent.  If connectivity is lost between the client and the broker during such a period, the client may not be able to detect that its published messages aren't reaching the broker.  This change introduces a new option ('fastClientDisconnectDetect') to *not* reschedule PINGREQ messages when sending packets.

This is in accordance with the MQTT spec v3.1.1, paragraph 3.1.2.10 (Keep Alive), which states that clients may send PINGREQ at any time:
```
The Client can send PINGREQ at any time, irrespective of the Keep Alive value, and use the PINGRESP to determine that the network and the Server are working.
```