Replace `pubrel` with `pubcomp`.
In QoS2 sequence,
when the client receives `publish`, the messege is stored into
`incomingStore` and sends pack `pubrec`. Then the server sends `pubrel`
to the client. Finally,  the client calls `handleMessage` with stored
message. If no errors are hannpened, `pubcomp` will be sent, but any errors
are happened, `pubcomp` shouldn't be sent.
This test is for the error case, so I replaced the description.