Supported `pause()` and `resume()` for `Store`'s stream.
If `pause()` and `resume()` can be used in `Store` implementation, it
is more flexible.
For example, `mqtt-level-store` can support preserving publish order
using `pause()` and `resume()`.

https://github.com/mcollina/mqtt-level-store

Here are the documentation of `pause()` and `resume()`.

https://nodejs.org/api/stream.html#stream_readable_pause
https://nodejs.org/api/stream.html#stream_readable_resume

According to the document, `pause()` and `resume()` work well with
'data' event, but not 'readable' event. So I replaced 'readable' with 'data'.