fix(connect/ws): ensure proxy stream is writable (#2024)
Here the proxy stream is being written, so we need to check if it is
writable ("which means the stream has not been destroyed, errored, or
ended, see
https://nodejs.org/dist/v18.19.0/docs/api/stream.html#writablewritable
).

Fixes #1914

I could reproduce the 'stream.push() after EOF' error in an app, but
unfortunately have no idea how to use it for a reproduction. Hopefully
this patch provides enough inspiration to write a test.